refactored assemblers and exceptions

+++ src/main/java/com/affablebean/exception/CustomerNotFoundException.java
@@ -7,7 +7,7 @@ public class CustomerNotFoundException extends RuntimeException {
 	 */
 	private static final long serialVersionUID = 1559360766655684177L;
 
-	public CustomerNotFoundException(Long id) {
+	public CustomerNotFoundException(Integer id) {
 		super("Could not find customer order " + id);
 	}
 }
