+++ src/main/java/com/affablebean/exception/CustomerOrderNotFoundException.java
@@ -7,7 +7,7 @@ public class CustomerOrderNotFoundException extends RuntimeException {
 	 */
 	private static final long serialVersionUID = 1559360766655684177L;
 
-	public CustomerOrderNotFoundException(Long id) {
+	public CustomerOrderNotFoundException(Integer id) {
 		super("Could not find customer " + id);
 	}
 }
