+++ src/main/java/com/affablebean/exception/MsgFeedbackNotFoundException.java
@@ -7,7 +7,7 @@ public class MsgFeedbackNotFoundException extends RuntimeException {
 	 */
 	private static final long serialVersionUID = -4911704767183845004L;
 
-	public MsgFeedbackNotFoundException(Long id) {
+	public MsgFeedbackNotFoundException(Integer id) {
 		super("Could not find message feedback " + id);
 	}
 }
