+++ src/main/java/com/affablebean/exception/MsgSubjectNotFoundException.java
@@ -7,7 +7,7 @@ public class MsgSubjectNotFoundException extends RuntimeException {
 	 */
 	private static final long serialVersionUID = -2261009936761844960L;
 
-	public MsgSubjectNotFoundException(Long id) {
+	public MsgSubjectNotFoundException(Integer id) {
 		super("Could not find message subject " + id);
 	}
 }
