+++ src/main/java/com/affablebean/exception/ProductNotFoundException.java
@@ -7,7 +7,7 @@ public class ProductNotFoundException extends RuntimeException {
 	 */
 	private static final long serialVersionUID = -3896146139774939779L;
 
-	public ProductNotFoundException(Long id) {
+	public ProductNotFoundException(Integer id) {
 		super("Could not find product " + id);
 	}
 }
