+++ src/main/java/com/affablebean/exception/PromotionNotFoundException.java
@@ -7,7 +7,7 @@ public class PromotionNotFoundException extends RuntimeException {
 	 */
 	private static final long serialVersionUID = -7225191521158738044L;
 
-	public PromotionNotFoundException(Long id) {
+	public PromotionNotFoundException(Integer id) {
 		super("Could not find promotion " + id);
 	}
 }
